requirements.txt
colorama==0.4.4
PyYAML==5.3.1
SQLAlchemy==1.3.22
websockets==8.1